adds mongodb connection to flask
