+++ flask/requirements.txt
@@ -1,4 +1,9 @@
-Flask
-python-dotenv
-langchain
-azure-ai-textanalytics
+blinker==1.9.0
+click==8.1.8
+dnspython==2.7.0
+Flask==3.1.0
+itsdangerous==2.2.0
+Jinja2==3.1.5
+MarkupSafe==3.0.2
+pymongo==4.11.1
+Werkzeug==3.1.3
